+++ py/desiutil/test/test_depend.py
@@ -173,5 +173,9 @@ class TestDepend(unittest.TestCase):
         self.assertTrue(getdep(hdr, 'sys').startswith('unknown'))
 
 
-if __name__ == '__main__':
-    unittest.main()
+def test_suite():
+    """Allows testing of only this module with the command::
+
+        python setup.py test -m <modulename>
+    """
+    return unittest.defaultTestLoader.loadTestsFromName(__name__)