+++ py/desiutil/test/test_install.py
@@ -395,5 +395,9 @@ class TestInstall(unittest.TestCase):
         self.assertFalse(isdir(self.desiInstall.working_dir))
 
 
-if __name__ == '__main__':
-    unittest.main()
+def test_suite():
+    """Allows testing of only this module with the command::
+
+        python setup.py test -m <modulename>
+    """
+    return unittest.defaultTestLoader.loadTestsFromName(__name__)