+++ py/desiutil/test/test_io.py
@@ -176,5 +176,9 @@ class TestIO(unittest.TestCase):
         self.assertEqual(dict2, {'a': {'b': {'p':3, 'q':4}}})
 
 
-if __name__ == '__main__':
-    unittest.main()
+def test_suite():
+    """Allows testing of only this module with the command::
+
+        python setup.py test -m <modulename>
+    """
+    return unittest.defaultTestLoader.loadTestsFromName(__name__)