+++ py/desiutil/test/test_top_level.py
@@ -38,5 +38,10 @@ class TestTopLevel(unittest.TestCase):
         else:
             self.assertRegexpMatches(theVersion, self.versionre)
 
-if __name__ == '__main__':
-    unittest.main()
+
+def test_suite():
+    """Allows testing of only this module with the command::
+
+        python setup.py test -m <modulename>
+    """
+    return unittest.defaultTestLoader.loadTestsFromName(__name__)