+++ doc/changes.rst
@@ -8,9 +8,11 @@ Change Log
 * Shorten Python version printed in dependency headers.
 * :mod:`desiutil.test.test_plots` was not cleaning up after itself.
 * Support new DESI+Anaconda software stack infrastructure (`PR #43`_).
+* fixes :meth:`~desiutil.bitmask.BitMask.names` when mask is a :class:`numpy.uint64` (`desitarget #79`_).
+* :meth:`~desiutil.bitmask.BitMask.names` is much faster.
 
 .. _`PR #43`: https://github.com/desihub/desiutil/pull/43
-
+.. _`desitarget #79`: https://github.com/desihub/desiutil/pull/41
 
 1.8.0 (2016-09-10)
 ------------------